pyproject.toml
[build-system]
requires = [
  "setuptools >= 40.9.0",
  "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "audio_feeder"
authors = [{name = "Paul Ganssle"}]
description = """
audio-feeder provides a server that serves your audiobooks and other audio
content as RSS feeds, with rich metadata, and provides a web frontent for navigation."""
license = {file = "LICENSE"}
requires-python = ">= 3.9"
dependencies = [
    "Flask>=0.11.1",
    "pyyaml>=6.0",
    "qrcode>=5.3",
    "Pillow>=3.4.2",
    "requests>=2.12.4",
    "jinja2",
    "click>=6.0",
    "progressbar2",
    "lxml>=2.0",
    "SQLAlchemy>=1.4",
    "attrs>=22.0",
    "typing_extensions>=4.0.0;python_version<'3.11'",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Framework :: Flask",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dynamic = ["version"]

[project.scripts]
audio-feeder = "audio_feeder.cli:cli"

[tool.setuptools]
packages = {find = {where = ["src"]}}
package-data = { audio_feeder=["data/**/*"] }

[tool.setuptools.dynamic]
version = {attr = "audio_feeder.__version__.VERSION"}

[tool.isort]
atomic=true
force_grid_wrap=0
include_trailing_comma=true
known_first_party = ["audio_feeder"]
known_third_party=[
    "click",
    "flask",
    "hypothesis",
    "jinja2",
    "lxml",
    "ruamel",
    "PIL",
    "progressbar2",
    "pytest",
    "pytz",
    "qrcode",
    "requests"
]
multi_line_output=3
use_parentheses=true
